feat(storage): fix PR review comments
